# backend/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: API del Sistema SAC de Clasificación de Solicitudes Académicas
  description: |
    API REST para el proyecto universitario SAC (Sistema de Atención y Clasificación).
    Centraliza las solicitudes académicas, las clasifica y prioriza, asigna personal,
    rastrea todo el ciclo de vida y mantiene un historial auditable.

    ## Autenticación
    Todos los endpoints (excepto login) requieren un token JWT Bearer válido.

    ## Roles
    | Rol            | Permisos                           |
    |----------------|------------------------------------|
    | ADMINISTRADOR  | Acceso total a todos los recursos  |
    | GESTOR         | Gestionar solicitudes y asignaciones |
    | CONSULTOR      | Acceso de solo lectura a solicitudes |

    ## Ciclo de Vida de Solicitudes
    ```
    REGISTRADA → CLASIFICADA → EN_ATENCION → ATENDIDA → CERRADA
    ```
    Cualquier otra transición retorna `400 Bad Request`. Las solicitudes cerradas no pueden modificarse.

    ## Reglas de Prioridad

    | Condición                              | Prioridad |
    |----------------------------------------|-----------|
    | CANCELACION cerca de fecha límite       | ALTA      |
    | HOMOLOGACION                           | ALTA      |
    | REGISTRO durante período de matrícula   | ALTA      |
    | CUPOS sin cupos disponibles            | ALTA      |
    | CONSULTA                               | BAJA      |
    | Por defecto                            | MEDIA     |

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/solicitudes:
    post:
      tags: [Solicitudes]
      summary: Crear una nueva solicitud académica
      description: |
        Registra una nueva solicitud académica en el sistema con estado `REGISTRADA`.
        **Rol requerido:** `GESTOR`
      operationId: crearSolicitud
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearSolicitudBody'
            example:
              tipo: CANCELACION
              descripcion: 'Estudiante solicita cancelación de Cálculo II por motivos médicos'
              canalOrigen: SAC
              idSolicitante: '1094000001'
      responses:
        '201':
          description: Solicitud creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
              example:
                id: 1
                tipo: CANCELACION
                descripcion: 'Estudiante solicita cancelación de Cálculo II por motivos médicos'
                canalOrigen: SAC
                fechaHoraRegistro: '2026-03-01T10:30:00Z'
                idSolicitante: '1094000001'
                prioridad: null
                justificacionPrioridad: null
                estado: REGISTRADA
                observacionCierre: null
        '400':
          $ref: '#/components/responses/SolicitudIncorrecta'
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '403':
          $ref: '#/components/responses/Prohibido'
        '500':
          $ref: '#/components/responses/ErrorInterno'

    get:
      tags: [Solicitudes]
      summary: Listar solicitudes con filtros y paginación
      description: |
        Retorna una lista paginada de solicitudes académicas. Soporta filtrado por estado,
        tipo, prioridad y usuario asignado.
        **Rol requerido:** `CONSULTOR`
      operationId: listarSolicitudes
      security:
        - BearerAuth: []
      parameters:
        - name: estado
          in: query
          required: false
          description: Filtrar por estado de solicitud
          schema:
            $ref: '#/components/schemas/EstadoSolicitud'
        - name: tipo
          in: query
          required: false
          description: Filtrar por tipo de solicitud
          schema:
            $ref: '#/components/schemas/TipoSolicitud'
        - name: prioridad
          in: query
          required: false
          description: Filtrar por nivel de prioridad
          schema:
            $ref: '#/components/schemas/Prioridad'
        - name: responsable
          in: query
          required: false
          description: Filtrar por ID del usuario asignado
          schema:
            type: integer
            format: int64
            example: 5
        - name: page
          in: query
          required: false
          description: Número de página (basado en cero)
          schema:
            type: integer
            default: 0
            minimum: 0
            example: 0
        - name: size
          in: query
          required: false
          description: Número de elementos por página
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
            example: 20
        - name: sort
          in: query
          required: false
          description: 'Campo y dirección de ordenamiento (ej. fechaHoraRegistro,desc)'
          schema:
            type: string
            example: 'fechaHoraRegistro,desc'
      responses:
        '200':
          description: Lista paginada de solicitudes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudesPaginadasResponse'
              example:
                content:
                  - id: 1
                    tipo: CANCELACION
                    descripcion: 'Cancelación de Cálculo II'
                    canalOrigen: SAC
                    fechaHoraRegistro: '2026-03-01T10:30:00Z'
                    idSolicitante: '1094000001'
                    prioridad: ALTA
                    justificacionPrioridad: 'CANCELACION cerca de fecha límite'
                    estado: CLASIFICADA
                    observacionCierre: null
                  - id: 2
                    tipo: CONSULTA
                    descripcion: 'Consulta sobre horario académico'
                    canalOrigen: CORREO
                    fechaHoraRegistro: '2026-03-01T11:00:00Z'
                    idSolicitante: '1094000002'
                    prioridad: BAJA
                    justificacionPrioridad: 'CONSULTA por defecto es BAJA'
                    estado: REGISTRADA
                    observacionCierre: null
                pagina: 0
                tamaño: 20
                totalElementos: 2
                totalPaginas: 1
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '403':
          $ref: '#/components/responses/Prohibido'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /api/v1/solicitudes/{id}:
    get:
      tags: [Solicitudes]
      summary: Obtener detalle de solicitud por ID
      description: |
        Retorna el detalle completo de una solicitud académica.
        **Rol requerido:** `CONSULTOR`
      operationId: obtenerSolicitudPorId
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdSolicitud'
      responses:
        '200':
          description: Detalle de la solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
              example:
                id: 1
                tipo: CANCELACION
                descripcion: 'Estudiante solicita cancelación de Cálculo II por motivos médicos'
                canalOrigen: SAC
                fechaHoraRegistro: '2026-03-01T10:30:00Z'
                idSolicitante: '1094000001'
                prioridad: ALTA
                justificacionPrioridad: 'CANCELACION cerca de fecha límite'
                estado: CLASIFICADA
                observacionCierre: null
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '403':
          $ref: '#/components/responses/Prohibido'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /api/v1/solicitudes/{id}/clasificar:
    patch:
      tags: [Solicitudes]
      summary: Clasificar y priorizar una solicitud
      description: |
        Establece el tipo, prioridad y justificación de prioridad para una solicitud.
        Cambia el estado a `CLASIFICADA`.
        **Rol requerido:** `GESTOR`
      operationId: clasificarSolicitud
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClasificarSolicitudBody'
            example:
              tipo: CANCELACION
              prioridad: ALTA
              justificacionPrioridad: 'CANCELACION cerca de fecha límite – requiere atención inmediata'
      responses:
        '200':
          description: Solicitud clasificada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
              example:
                id: 1
                tipo: CANCELACION
                descripcion: 'Estudiante solicita cancelación de Cálculo II por motivos médicos'
                canalOrigen: SAC
                fechaHoraRegistro: '2026-03-01T10:30:00Z'
                idSolicitante: '1094000001'
                prioridad: ALTA
                justificacionPrioridad: 'CANCELACION cerca de fecha límite – requiere atención inmediata'
                estado: CLASIFICADA
                observacionCierre: null
        '400':
          $ref: '#/components/responses/SolicitudIncorrecta'
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '403':
          $ref: '#/components/responses/Prohibido'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /api/v1/solicitudes/{id}/estado:
    patch:
      tags: [Solicitudes]
      summary: Cambiar estado de solicitud
      description: |
        Transiciona una solicitud a un nuevo estado siguiendo el ciclo de vida válido:
        `REGISTRADA → CLASIFICADA → EN_ATENCION → ATENDIDA → CERRADA`.
        Cualquier transición inválida retorna `400 Bad Request`.
        **Rol requerido:** `GESTOR`
      operationId: cambiarEstadoSolicitud
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambiarEstadoBody'
            example:
              nuevoEstado: EN_ATENCION
              observacion: 'La solicitud está siendo atendida por el personal asignado'
      responses:
        '200':
          description: Estado cambiado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
              example:
                id: 1
                tipo: CANCELACION
                descripcion: 'Estudiante solicita cancelación de Cálculo II por motivos médicos'
                canalOrigen: SAC
                fechaHoraRegistro: '2026-03-01T10:30:00Z'
                idSolicitante: '1094000001'
                prioridad: ALTA
                justificacionPrioridad: 'CANCELACION cerca de fecha límite'
                estado: EN_ATENCION
                observacionCierre: null
        '400':
          description: Transición de estado inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2026-03-01T12:00:00Z'
                status: 400
                error: Solicitud Incorrecta
                message: 'Transición de estado inválida de REGISTRADA a EN_ATENCION'
                path: /api/v1/solicitudes/1/estado
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '403':
          $ref: '#/components/responses/Prohibido'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /api/v1/solicitudes/{id}/asignar:
    post:
      tags: [Solicitudes]
      summary: Asignar un usuario responsable a una solicitud
      description: |
        Asigna un usuario como la persona responsable de una solicitud.
        El usuario debe estar activo.
        **Rol requerido:** `GESTOR`
      operationId: asignarResponsable
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsignarResponsableBody'
            example:
              usuarioId: 5
      responses:
        '201':
          description: Usuario asignado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsignacionResponse'
              example:
                id: 10
                solicitudId: 1
                usuarioId: 5
                fechaAsignacion: '2026-03-01T14:00:00Z'
                activa: true
        '400':
          $ref: '#/components/responses/SolicitudIncorrecta'
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '403':
          $ref: '#/components/responses/Prohibido'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /api/v1/solicitudes/{id}/cerrar:
    patch:
      tags: [Solicitudes]
      summary: Cerrar una solicitud
      description: |
        Cierra una solicitud. La solicitud debe estar en estado `ATENDIDA`.
        Se requiere una observación de cierre. Una vez cerrada, la solicitud no puede modificarse.
        **Rol requerido:** `GESTOR`
      operationId: cerrarSolicitud
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdSolicitud'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CerrarSolicitudBody'
            example:
              observacionCierre: 'Solicitud resuelta – cancelación aprobada por el comité académico'
      responses:
        '200':
          description: Solicitud cerrada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
              example:
                id: 1
                tipo: CANCELACION
                descripcion: 'Estudiante solicita cancelación de Cálculo II por motivos médicos'
                canalOrigen: SAC
                fechaHoraRegistro: '2026-03-01T10:30:00Z'
                idSolicitante: '1094000001'
                prioridad: ALTA
                justificacionPrioridad: 'CANCELACION cerca de fecha límite'
                estado: CERRADA
                observacionCierre: 'Solicitud resuelta – cancelación aprobada por el comité académico'
        '400':
          description: La solicitud no está en estado ATENDIDA o falta observación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: '2026-03-01T16:00:00Z'
                status: 400
                error: Solicitud Incorrecta
                message: 'No se puede cerrar la solicitud – estado actual es EN_ATENCION, se esperaba ATENDIDA'
                path: /api/v1/solicitudes/1/cerrar
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '403':
          $ref: '#/components/responses/Prohibido'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '500':
          $ref: '#/components/responses/ErrorInterno'
  /api/v1/solicitudes/{id}/historial:
    get:
      tags: [Solicitudes]
      summary: Obtener historial completo de auditoría de una solicitud
      description: |
        Retorna el rastro de auditoría completo de todas las acciones realizadas en una solicitud,
        ordenado cronológicamente.
        **Rol requerido:** `CONSULTOR`
      operationId: obtenerHistorialSolicitud
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdSolicitud'
      responses:
        '200':
          description: Lista de entradas del historial
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistorialResponse'
              example:
                - id: 1
                  fechaHora: '2026-03-01T10:30:00Z'
                  accion: 'REGISTRO'
                  usuarioResponsable: 'jgarcia'
                  observaciones: 'Solicitud creada vía portal SAC'
                - id: 2
                  fechaHora: '2026-03-01T11:00:00Z'
                  accion: 'CLASIFICACION'
                  usuarioResponsable: 'jgarcia'
                  observaciones: 'Clasificada como CANCELACION con prioridad ALTA'
                - id: 3
                  fechaHora: '2026-03-01T14:00:00Z'
                  accion: 'ASIGNACION'
                  usuarioResponsable: 'mrodriguez'
                  observaciones: 'Asignada a mrodriguez para gestión'
        '401':
          $ref: '#/components/responses/NoAutorizado'
        '403':
          $ref: '#/components/responses/Prohibido'
        '404':
          $ref: '#/components/responses/NoEncontrado'
        '500':
          $ref: '#/components/responses/ErrorInterno'

  # ─────────────────────────────────────────────
  #  USUARIOS
  # ─────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido de POST /api/v1/usuarios/login

  # ── Parámetros Reutilizables ──────────────────
  parameters:
    IdSolicitud:
      name: id
      in: path
      required: true
      description: ID único de solicitud
      schema:
        type: integer
        format: int64
        example: 1

  # ── Esquemas ──────────────────────────────────
  schemas:
...
    TipoSolicitud:
      type: string
      description: Tipo de solicitud académica
      enum:
        - REGISTRO
        - HOMOLOGACION
        - CANCELACION
        - CUPOS
        - CONSULTA
    CanalOrigen:
      type: string
      description: Canal de origen por el cual se recibió la solicitud
      enum:
        - CSU
        - EMAIL
        - WEB
        - SAC
        - TELEFONICO
        - PRESENCIAL
    Prioridad:
      type: string
      description: Nivel de prioridad asignado a la solicitud
      enum:
        - ALTA
        - MEDIA
        - BAJA
    EstadoSolicitud:
      type: string
      description: |
        Estado actual en el ciclo de vida de la solicitud.
        Transiciones válidas: REGISTRADA → CLASIFICADA → EN_ATENCION → ATENDIDA → CERRADA
      enum:
        - REGISTRADA
        - CLASIFICADA
        - EN_ATENCION
        - ATENDIDA
        - CERRADA
...
    CrearSolicitudBody:
      type: object
      description: Payload para crear una nueva solicitud académica
      required:
        - estudianteNombre
        - estudianteCorreo
        - estudianteTelefono
        - estudianteIdentificacion
        - asunto
        - descripcion
        - canalOrigen
      properties:
        estudianteNombre:
          type: string
          minLength: 2
          maxLength: 100
          description: Nombre completo del estudiante
          example: 'María González'
        estudianteCorreo:
          type: string
          format: email
          maxLength: 100
          description: Correo electrónico del estudiante
          example: 'maria.gonzalez@uniquindio.edu.co'
        estudianteTelefono:
          type: string
          minLength: 7
          maxLength: 20
          description: Número de teléfono del estudiante
          example: '3001234567'
        estudianteIdentificacion:
          type: string
          minLength: 1
          maxLength: 50
          description: Número de identificación del estudiante
          example: '1094567890'
        asunto:
          type: string
          minLength: 5
          maxLength: 200
          description: Asunto o título breve de la solicitud
          example: 'Solicitud de homologación de materias'
        descripcion:
          type: string
          minLength: 10
          maxLength: 2000
          description: Descripción detallada de la solicitud académica
          example: 'Deseo homologar las materias Cálculo I y Álgebra Lineal que cursé en mi universidad anterior'
        canalOrigen:
          $ref: '#/components/schemas/CanalOrigen'
    ClasificarSolicitudBody:
      type: object
      description: Payload para clasificar y establecer prioridad de una solicitud
      required:
        - tipo
        - prioridad
        - notaClasificacion
      properties:
        tipo:
          $ref: '#/components/schemas/TipoSolicitud'
        prioridad:
          $ref: '#/components/schemas/Prioridad'
        notaClasificacion:
          type: string
          minLength: 5
          maxLength: 500
          description: Nota o justificación para la clasificación y prioridad asignada
          example: 'Solicitud de homologación con documentación completa'
    CambiarEstadoBody:
      type: object
      description: Payload para transicionar una solicitud a un nuevo estado
      required:
        - nuevoEstado
      properties:
        nuevoEstado:
          $ref: '#/components/schemas/EstadoSolicitud'
        nota:
          type: string
          maxLength: 1000
          description: Nota opcional sobre el cambio de estado
          example: 'Iniciando atención de la solicitud'
    AsignarResponsableBody:
      type: object
      description: Payload para asignar un usuario responsable a una solicitud
      required:
        - responsableId
      properties:
        responsableId:
          type: integer
          format: int64
          description: ID del usuario a asignar como responsable
          example: 2
        notaAsignacion:
          type: string
          maxLength: 500
          description: Nota opcional sobre la asignación
          example: 'Asignación por experiencia en homologaciones'
    CerrarSolicitudBody:
      type: object
      description: Payload para cerrar una solicitud (requiere resolución y notas de cierre)
      required:
        - resolucion
      properties:
        resolucion:
          type: string
          minLength: 10
          maxLength: 2000
          description: Resolución obligatoria explicando cómo se resolvió la solicitud
          example: 'Se aprobó la homologación de las materias Cálculo I y Álgebra Lineal. El estudiante debe presentar los certificados en la oficina de registro.'
        notasCierre:
          type: string
          maxLength: 1000
          description: Notas adicionales opcionales sobre el cierre
          example: 'Solicitud resuelta satisfactoriamente'
...
    SolicitudResponse:
      type: object
      description: Representación completa de una solicitud académica
      properties:
        id:
          type: integer
          format: int64
          description: ID único de la solicitud
          example: 1
        estudianteNombre:
          type: string
          description: Nombre completo del estudiante
          example: 'María González'
        estudianteCorreo:
          type: string
          description: Correo electrónico del estudiante
          example: 'maria.gonzalez@uniquindio.edu.co'
        estudianteTelefono:
          type: string
          description: Número de teléfono del estudiante
          example: '3001234567'
        estudianteIdentificacion:
          type: string
          description: Número de identificación del estudiante
          example: '1094567890'
        asunto:
          type: string
          description: Asunto o título de la solicitud
          example: 'Solicitud de homologación de materias'
        descripcion:
          type: string
          description: Descripción detallada de la solicitud
          example: 'Deseo homologar las materias Cálculo I y Álgebra Lineal que cursé en mi universidad anterior'
        canalOrigen:
          $ref: '#/components/schemas/CanalOrigen'
        fechaHoraRegistro:
          type: string
          format: date-time
          description: Timestamp cuando fue registrada la solicitud
          example: '2026-03-01T10:30:00Z'
        tipo:
          $ref: '#/components/schemas/TipoSolicitud'
          nullable: true
        prioridad:
          $ref: '#/components/schemas/Prioridad'
          nullable: true
        notaClasificacion:
          type: string
          nullable: true
          description: Nota de clasificación y justificación de la prioridad
          example: 'Solicitud de homologación con documentación completa'
        estado:
          $ref: '#/components/schemas/EstadoSolicitud'
        resolucion:
          type: string
          nullable: true
          description: Resolución de la solicitud (solo presente cuando estado es CERRADA)
          example: null
        notasCierre:
          type: string
          nullable: true
          description: Notas adicionales del cierre (solo presente cuando estado es CERRADA)
          example: null
    SolicitudesPaginadasResponse:
      type: object
      description: Wrapper de respuesta paginada para solicitudes
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SolicitudResponse'
        pagina:
          type: integer
          description: Número de página actual (basado en cero)
          example: 0
        tamaño:
          type: integer
          description: Número de elementos por página
          example: 20
        totalElementos:
          type: integer
          format: int64
          description: Número total de registros que coinciden
          example: 42
        totalPaginas:
          type: integer
          description: Número total de páginas
          example: 3
...
    AsignacionResponse:
      type: object
      description: Representación de una asignación solicitud-a-usuario
      properties:
        id:
          type: integer
          format: int64
          description: ID único de la asignación
          example: 10
        solicitudId:
          type: integer
          format: int64
          description: ID de la solicitud asignada
          example: 1
        usuarioId:
          type: integer
          format: int64
          description: ID del usuario asignado
          example: 5
        fechaAsignacion:
          type: string
          format: date-time
          description: Timestamp de la asignación
          example: '2026-03-01T14:00:00Z'
        activa:
          type: boolean
          description: Si esta asignación está actualmente activa
          example: true
...
    ErrorResponse:
      type: object
      description: Cuerpo de respuesta de error estándar
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp de ocurrencia del error
          example: '2026-03-01T12:00:00Z'
        status:
          type: integer
          description: Código de estado HTTP
          example: 400
        error:
          type: string
          description: Frase de razón del error HTTP
          example: 'Solicitud Incorrecta'
        message:
          type: string
          description: Mensaje de error detallado
          example: 'Transición de estado inválida de REGISTRADA a EN_ATENCION'
        path:
          type: string
          description: Ruta de solicitud que causó el error
          example: '/api/v1/solicitudes/1/estado'

    # ── Esquemas de IA (Opcionales – RF-09, RF-10, RF-11) ─────────────
...
  responses:
    SolicitudIncorrecta:
      description: Solicitud Incorrecta – Entrada inválida o violación de regla de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-03-01T12:00:00Z'
            status: 400
            error: Solicitud Incorrecta
            message: 'Falla en validación: descripcion no debe estar en blanco'
            path: /api/v1/solicitudes
    NoAutorizado:
      description: No Autorizado – Token JWT faltante o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-03-01T12:00:00Z'
            status: 401
            error: No Autorizado
            message: 'Se requiere autenticación completa para acceder a este recurso'
            path: /api/v1/solicitudes
    Prohibido:
      description: Prohibido – Permisos de rol insuficientes
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-03-01T12:00:00Z'
            status: 403
            error: Prohibido
            message: 'Acceso denegado – rol CONSULTOR no puede realizar esta acción'
            path: /api/v1/solicitudes
    NoEncontrado:
      description: No Encontrado – El recurso no existe
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-03-01T12:00:00Z'
            status: 404
            error: No Encontrado
            message: 'Solicitud con ID 999 no encontrada'
            path: /api/v1/solicitudes/999
    ErrorInterno:
      description: Error Interno del Servidor – Falla inesperada
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            timestamp: '2026-03-01T12:00:00Z'
            status: 500
            error: Error Interno del Servidor
            message: 'Ocurrió un error inesperado. Por favor intente de nuevo más tarde.'
            path: /api/v1/solicitudes